# Don't convert chunks to matrixes and then merges them...