--- a/pkg/querier/chunk_queryable.go
+++ b/pkg/querier/chunk_queryable.go
@@ -16,9 +16,12 @@ type ChunkStore interface {
 }
 
 // NewQueryable creates a new Queryable for cortex.
-func NewQueryable(distributor Distributor, chunkStore ChunkStore) storage.Queryable {
+func NewQueryable(distributor Distributor, chunkStore ChunkStore, iterators bool) storage.Queryable {
 	dq := newDistributorQueryable(distributor)
 	cq := newChunkQueryable(chunkStore)
+	if iterators {
+		cq = newIterChunkQueryable(chunkStore)
+	}
 
 	return storage.QueryableFunc(func(ctx context.Context, mint, maxt int64) (storage.Querier, error) {
 		dqr, err := dq.Querier(ctx, mint, maxt)