...use iterators instead. 

Also:
- Update remote read endpoint for SeriesSet interface.
- Copy the upstream heap-based merging code, and update it for chunk boundaries.
- Dedupe samples coming out of the merge iterator.

Big drawback of this approach is the optimisation necessary about `Seek` - as we don't know the end time for any particular-seek'd iterator, we aren't technically able to filter out chunks in the future.  This has a large performance impact (basically O(n^2)), as seeks at the very beginning of the range have to include all chunks.  To avoid this I've assumed an end time of 48hrs, meaning vector queries with duration larger than 24hrs won't work correctly (ie `sum(rate(foo[48h]))`).  But OTOH this allows code closer to O(n log n).

Before vs after:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkChunkQueryable-4     364386071     44484806      -87.79%

benchmark                     old allocs     new allocs     delta
BenchmarkChunkQueryable-4     1019223        453391         -55.52%

benchmark                     old bytes     new bytes     delta
BenchmarkChunkQueryable-4     903011968     14560358      -98.39%
```

Thats about 8.2x speed up in ns/op and 62x less memory usage.

Part of #209, closes #81 as redundant.