@tomwilkie 2018-02-20
We're seeing some more issues with this in testing.  Worth reviewing, but don't merge yet.