@tomwilkie 2018-07-10
Found an optimisation in the ChunksToMatrix function too, so results are now:

```
$ go test ./pkg/querier -benchmem -bench BenchmarkChunkQueryable 
goos: darwin
goarch: amd64
pkg: github.com/weaveworks/cortex/pkg/querier
BenchmarkChunkQueryable/matrixes/DoubleDelta-4                  100      20379796 ns/op    42817634 B/op       23292 allocs/op
BenchmarkChunkQueryable/iterators/DoubleDelta-4                  20      76533564 ns/op      984637 B/op       16846 allocs/op
BenchmarkChunkQueryable/matrixes/Varbit-4                       100      19467781 ns/op    42817585 B/op       22573 allocs/op
BenchmarkChunkQueryable/iterators/Varbit-4                       20      81691003 ns/op      984516 B/op       16125 allocs/op
PASS
ok      github.com/weaveworks/cortex/pkg/querier    8.490s
```

I suspect it might make sense to pick the strategy (iterate or merge) based on the length of the query (or the number of chunks in the series).