@tomwilkie 2018-07-10
Needed to update the vendored Prometheus for https://github.com/prometheus/prometheus/commit/78efdc6d