@tomwilkie 2018-07-13
> Are you thinking at some point use matrix for queries with a small number of chunks for speed, and iterator for a large number of chunks to not OOM?

Right now I'm just trying to stop the querier OOMing for high cardinality / long queries / lots of simultaneous queries.  But yeah, once I've got that working it probably makes a lot of sense to switch based on number of chunks per series - good idea!